Fix blocking in mvnc detctor due to queue blocked in put

+++ snpx/apps/mvnc_dev.py
@@ -29,8 +29,7 @@ try:
                 self.ncs_graph.SetGraphOption(mvnc.GraphOption.DONT_BLOCK, 1)
 
         def load_input(self, input_tensor):
-            input_tensor = input_tensor.astype(np.float16)
-            self.ncs_graph.LoadTensor(input_tensor, None)
+            self.ncs_graph.LoadTensor(input_tensor.astype(np.float16), None)
 
         def get_output(self):
             ncs_out = None
